Moved config types
