+++ src/types/Config.ts
@@ -1,7 +1,4 @@
-import { AnimationDisplayConfig } from '../animation/Display';
-import { ZoomConfig } from '../Zoom';
-import { AnimationConfig, CounterConfig } from './Animation';
-
+/** Global */
 export type LightboxConfig = {
     targetClass: string;
     animation?: {
@@ -12,16 +9,52 @@ export type LightboxConfig = {
     display?: Partial<DisplayConfig>;
 };
 
+/** Display */
 export type DisplayConfig = {
     rounded: boolean;
     animation?: Partial<AnimationDisplayConfig>;
 };
 
-export type UIConfig = {
-    animation: AnimationConfig;
-    zoom: ZoomConfig;
-    counter: CounterConfig;
+export type AnimationDisplayMorphConfig = {
+    enabled: boolean;
+    duration: number;
 };
+
+export type AnimationDisplayConfig = {
+    enabled: boolean;
+    duration: number;
+    next: string;
+    prev: string;
+    morph: Partial<AnimationDisplayMorphConfig>;
+};
+
+/** Counter */
+
+type CounterXPosition = 'left' | 'center' | 'right';
+type CounterYPosition = 'top' | 'bottom';
+
+export type CounterConfig = {
+    enabled: boolean;
+    y: CounterYPosition;
+    x: CounterXPosition;
+    animation?: Partial<AnimationCounterConfig>;
+};
+
+export type AnimationCounterConfig = {
+    enabled: boolean;
+    duration: number;
+    keyframes: string;
+};
+
+/** Zoom */
+
+export type ZoomConfig = {
+    enabled: boolean;
+    blockNative: boolean;
+    level: number;
+};
+
+/** Dock */
 
 export type DockConfig = {
     enabled: boolean;
